office change

+++ database/migrations/2024_09_18_101708_create_purchase_order_items_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('purchase_order_id')->references('id')->on('purchase_orders')->onDelete('cascade');
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->decimal('quantity', 8, 2);
+            $table->integer('quantity');
             $table->decimal('unit_price', 8, 2);
             $table->decimal('total_price', 8, 2);
             $table->decimal('tax', 10, 2)->default(0.00);
